+++ src/@ionic-native/plugins/open-native-settings/index.ts
@@ -1,11 +1,11 @@
-import { Plugin, IonicNativePlugin, Cordova } from '@ionic-native/core';
 import { Injectable } from '@angular/core';
+import { Cordova, IonicNativePlugin, Plugin } from '@ionic-native/core';
 
 /**
  * @name Open Native Settings
  * @description
- * Plugin to open native screens of iOS/android settings 
- * @usage 
+ * Plugin to open native screens of iOS/android settings
+ * @usage
  * You can open any of these settings:
  * ```
  * "about", // ios
@@ -78,7 +78,7 @@ import { Injectable } from '@angular/core';
     "wifi_ip", // android
     "wifi", // ios, android
     "wireless" // android
-    ``` 
+    ```
  * ```typescript
  * import { OpenNativeSettings } from '@ionic-native/open-native-settings';
  *
@@ -99,13 +99,13 @@ import { Injectable } from '@angular/core';
 })
 @Injectable()
 export class OpenNativeSettings extends IonicNativePlugin {
-
   /**
    * Opens a setting dialog
-   * @param setting {string} setting name
+   * @param setting {string|array} setting name
    * @return {Promise<any>}
    */
   @Cordova()
-  open(setting: string): Promise<any> { return; }
-
+  open(setting: string | [string, boolean]): Promise<any> {
+    return;
+  }
 }